pyproject.toml
[project]
name = "prompt-pulse"
version = "0.1.0"
description = "Voice-activated terminal-aware prompt enhancer for AI coding assistants"
readme = "SPEC.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.26",
    "webrtcvad>=2.0.10",
    "whispercpp>=0.0.17",
    "litellm>=1.30",
    "pynput>=1.7.6",
    "pyperclip>=1.8",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "typer[all]>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
iterm2 = [
    "iterm2>=2.7",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.3",
]

[project.scripts]
prompt-pulse = "prompt_pulse.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_pulse"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Shell hook scripts embedded as string literals contain long lines
"src/prompt_pulse/terminal/monitor.py" = ["E501"]
# Regex patterns are long by nature
"src/prompt_pulse/terminal/error_patterns.py" = ["E501"]
# Test strings contain long error messages
"tests/test_error_patterns.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
